M3.5a: split gateway_client into auth/meters/inverters submodules + magic-literal sweep
- Move gateway_client.rs → gateway_client/{mod,auth,meters,inverters}.rs
- auth.rs: check_jwt() + parse_session_cookie()
- meters.rs: probe_meters(), get_meter_readings(), extract_cumulatives_from_json()
- inverters.rs: get_inverter_snapshots()
- mod.rs: struct, new(), cookie_header(), auth_header() helper, parse_session_cookie()
- Promote HTTP_TIMEOUT_SECS, NET_CONSUMPTION_MEASUREMENT_TYPE, METER_STATE_ENABLED to named consts
- Replace repeated "Authorization"/"Bearer" strings with auth_header() helper
- Replace repeated "net-consumption"/"enabled" strings with consts
- Update scheduler/mod.rs import paths
